fixed issue with range() call

+++ bayesian_bootstrap/bootstrap.py
@@ -43,10 +43,10 @@ def central_credible_interval(samples, alpha=0.05):
 
 def highest_density_interval(samples, alpha=0.05):
     samples_sorted = sorted(samples)
-    window_size = len(samples) - round(len(samples)*alpha)
+    window_size = int(len(samples) - round(len(samples)*alpha))
     smallest_window = (None, None)
     smallest_window_length = float('inf')
-    for i in range(int(len(samples_sorted)) - window_size):
+    for i in range(len(samples_sorted) - window_size):
         window = samples_sorted[i+window_size-1], samples_sorted[i]
         window_length = samples_sorted[i+window_size-1] - samples_sorted[i]
         if window_length < smallest_window_length:
